cast duplicate count to double before calculating percentage

+++ qprofiler/src/org/qcmg/qprofiler/bam/BamSummaryReport.java
@@ -327,7 +327,7 @@ public class BamSummaryReport extends SummaryReport {
 		Element noOfReadsE = createSubElement(summaryElement, "ReadCount");
 		noOfReadsE.setAttribute("value", noOfRecords + "");
 		Element noOfDupsE = createSubElement(summaryElement, "DuplicatePercentage");
-		noOfDupsE.setAttribute("value", (duplicateCount / noOfRecords) * 100 + "");
+		noOfDupsE.setAttribute("value", ((double)duplicateCount / noOfRecords) * 100 + "");
 		
 		// loop through flag, tallying unmapped
 		long tally = 0;
